+++ src/ambition/version_details.cc
@@ -24,6 +24,7 @@ std::vector<VersionUpdates> versionDetails = {
           _("Show an information panel for Forts, with the general or king leadership level, number of soldiers, and the approximate total strength, including the effects of the leader."),
           _("Show information panels for other buildings, with information relevant to them, such as worker and good counts."),
           _("Show an information panel for Fryhtan Lairs with their species name and level.  The level is an approximate relative reflection of the species' strength and the gold reward for killing Ordos.  Note that this reflects only the species, not the current unit count and combat level, which can only be seen with a Phoenix."),
+          _("Added a progress bar for Seats of Power."),
         },
       },
       {